Add pac-ret hardening to _init and _fini (#4)
Ideally, this should be conditional on whether ptrauth_returns is
requested and which key is used, but this patch still should be safe as
PACIASP and AUTIASP are encoded as HINT and both prologue and epilogue
use the same IA key.

Note that even if .init_array and .fini_array are actually used,
_init and _fini functions are statically linked into every executable,
thus this patch is a natural way to silence multiple warnings reported
by PAuth gadget scanner for every executable.